requirements.txt
Django
South
argparse
defusedxml
django-filter
django-oauth-plus
django-oauth2-provider
djangorestframework
gunicorn
oauth2
pymongo
shortuuid
